part6: ex6.7
